Add note and backup download
