+++ package.json
@@ -16,7 +16,6 @@
     "@quasar/extras": "^1.0.0",
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
-    "js-file-download": "^0.4.12",
     "mitt": "^2.1.0",
     "quasar": "^2.0.0-rc.2",
     "vue-masonry": "^0.13.0"
